Fixed imports and types 😂

+++ utils.py
@@ -6,12 +6,13 @@ logging_level = logging.DEBUG
 log_format = '%(asctime)s [%(name)s] %(message)s'
 log_date_format = '%d-%m-%Y %HL%M:%S'
 
+
 def get_x_and_y_offset(boundaries: tuple[int, int, int, int]) -> Tuple[float, float]:
     game_corners = [x/2 for x in boundaries]
     return game_corners[0], game_corners[1]
 
 
-def get_width_and_height(boundaries: tuple[int, int, int, int]) -> Tuple[int, int]:
+def get_width_and_height(boundaries: tuple[int, int, int, int]) -> Tuple[float, float]:
     game_corners = [x/2 for x in boundaries]
     width = game_corners[2] - game_corners[0]
     height = game_corners[3] - game_corners[1]
